Add uuid-generation to ensure all messages get id'ed.
In order for clients to rely on message ids existing, we are updating
this client library to always add the id if one does not exist.

After discussion with the group, this method was decided on in favor of
* requiring all services to add ids themselves (too easy to mess
  up/forget and exposes ESB implementation details)
* having the ESB generate the ids itself (would require the ESB to
  respond to the service with the ID generated).

We also decided that the ESB itself should reject messages that come
through without ids.  This will allow the clients to rely on ids being
present on all messages.