Change the subscribe signature to indicate single groups.
This is a bit opinionated, but I doubt we will want to subscribe to
multiple different groups at one time (doing so requires filtering the
different channel messages in the client).  This doesn't actually remove
the ability to subscribe to multiple groups at once as all arguments
passed to subscribe are passed along, but it does hide that away as more
of a hack.

I want to update things here so that the return value of the subscribe
call is something that you can register events on to just get messages
from that subscription.  A second subscription (when supported) would
give you a different "channel" to receive events on.