Add the ability for rpc responses to be CC'd to requested groups.
This allows the requester to dictate where it expects a response without
the RPC service knowing anything particular about the requester.  For
example, a routing service can be created that only listens on a single
group but by using this mechanism, all messages it sends to the ESB can
be appropriately forwarded back to the router by adding the CC field.