Fix forgotten auth promise rejection.
Bluebird is a bit funny when it comes to these unhandled promises.  It
likes to dump stack traces very greedily, so while the code worked
before, it erroneously dumped an error all the time.  Therefore we need
to make sure we are only creating the rejected promise when we need it
to actually fail things.