Convert library from using data/meta pairs to a Message object.
Introduce a domain object for representing an ESB Message.  This breaks
compatibility with previous versions, primarily with the places where
messages are sent back to the calling code as it is no longer returning
the `data`, but the entire message in a message object.

When needed, a new MessageBuilder was created and is available at
`client.mb` to facilitate building the messages and linking them
properly for the ESB.

The new Message objects also expose functionality from
[object-path](https://github.com/mariocasciaro/object-path) to make them
easier to work with.

This change was made primarily for the purpose of providing an object
that we can add new functionality to.  POJO's are simple, but they bloat
the API of this library and make it harder to add more state-based
behaviors.