Add `inre` to send methods for ESB message tracing.
Like the docs say below, this provides a mechanism for tracing a single
call to some ESB service down into secondary/tertiary/n-ary services
that were made in order to fulfill the original request.

This provides an interesting capability to the ESB where we can more
easily tell the true weight of service calls, but it does require the
clients to utilize the field properly to work.

A preferred way of implementing this was to alter the rpcReceive
callback call to pass a proxy client that knew to pass the inre flag on
its otherwise-unmodified send/rpcSend methods.  This would have been a
clean way of doing things, but unfortunately there doesn't seem to be a
way to do this yet and our best hope is ecmascript 6 Proxy handlers.
Instead, this less ideal method was attempted that requires a new
parameter on the send/rpcSend calls.