Send rpc replies directly to the caller's channel id.
Right now, our services have to subscribe (automatically) to the
SERVICE-result groups.  This leads to popular RPC services having their
responses duplicated and sent to a ton of channels that don't actually
need them.

We do still want to have SERVICE-result groups, however, as they can be
useful for some cases (like the -sonar services).  We just need to
provide a way for RPC transactions to happen without having to subscribe
to the SERVICE-result groups.

We have actually had the mechanism we need for this for a while.
Channels are already uniquely identified by an id and send that id as
their "from" id in the meta.  We can also direct a message to that
channel by using its id in a "channel" id in the meta.

By adding the "channel" id, the calling service gets the message
directly and we can remove the "-result" subscription (assuming all
services follow this pattern).

One important note, is that if you do subscribe to the SERVICE-result
group and send rpc calls to the SERVICE group, the ESB will send the
reply back twice (once for the "channel" header and once for the "group"
header.  Assuming you used `rpcSend` to send the request, though, even
that may not matter as `rpcSend` and Promises use `.once` style behavior
that will only resolve a single time.