Add newrelic support for rpcReceive.
rpcReceive handlers are the only ones that can be properly wrapped in a
newrelic transaction.  Others would require custom implementation in the
calling code as the esb client can't know when the transaction ends.

This is currently using a background transaction as that is more
correct semantically.  I'm not positive the features in newrelic are as
nice for background transactions, however, so we may want to switch to a
web transaction in the future, even if that means hacking things into
place.

Errors will be sent to newrelic as well.  Included will be the original
message for help with debugging what happened.

The newrelic instrumentation is a bit weird.  There doesn't seem to be
support for bluebird promises in newrelic's agent and because of the
magicalness of newrelic's transaction tracing, there's not a lot we can
do about it.  It's possible to use newrelic.createTracer() around the
error handler and finally clause, but that seemed to double/treble up
the time taken in the transaction which is not what we want (it was
reporting time for the error handler even when it wasn't running).
Instead, I dig a little below newrelic's surface API and manually bind
the function to the active transaction segment.  I'm fairly certain this
should work, even in the presence of concurrent transactions, but I'm
not 100% positive.  This also has the downside of possibly breaking on
new versions of the newrelic agent.  We'll have to be careful when
upgrading the newrelic agent to make sure things continue working.