Don't debug internal messages for now.
Until we update the debug code in here, this serves to cut down on the
lines of debug being spit out and only show the important messages for
the actual service.

The `_isInternal` abstraction is my attempt to better define the
behavior.  It helps make this feel like less of a special case in the if
statements in `_send` and `_handlePacket`.

I needed to be able to deal with actual message objects to make this
work well, so I moved the debug around to occur at the right times.
This also meant that we needed a new log message for when we fail to
parse a message as it would never become a real message object.